requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
gtts>=2.4.0
Pillow>=10.0.0